% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{commons}
\alias{commons}
\alias{computePValueMain}
\alias{modelName}
\title{Common functions}
\usage{
computePValueMain(object, n, model = c("gam", "poly"), type = c("nc",
  "c", "ct"), d = NULL)

modelName(type, model, d)
}
\arguments{
\item{object}{Numeric value or an object (\code{fHTEST}, \code{ur.df} or
\code{htest}) for which p-value needs to
approximated.}

\item{n}{Sample size.}

\item{model}{The model type to be used for approximation. Available is GAM
and polynomial regression. If \code{gam} is chosen, then \code{d} has no
effect.}

\item{type}{The type of unit root test. Currently supports: \code{nc} for
test without drift and trend, \code{c} for test with only drift and
\code{ct} for test with both drift and trend.}

\item{d}{The degree for polynomial. \code{d} must be \eqn{\geq 3} and
\eqn{\leq 6}. If \code{gam} is chosen, then \code{d} has no effect.}
}
\description{
Common functions used by top level functions.
}
\details{
\code{modelName} generates the name of function based on the arguments
\code{model}, \code{type} and \code{d} (when \code{model = "poly"}).\cr
\code{computePValueMain} is called by the other top level functions to
approximate the p-value.
}
