% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\alias{woTD_poly3}
\alias{woTD_poly4}
\alias{woTD_poly5}
\alias{woTD_poly6}
\alias{woTD_gam}
\alias{wD_poly3}
\alias{wD_poly4}
\alias{wD_poly5}
\alias{wD_poly6}
\alias{wD_gam}
\alias{wTD_poly3}
\alias{wTD_poly4}
\alias{wTD_poly5}
\alias{wTD_poly6}
\alias{wTD_gam}
\title{Models}
\usage{
woTD_poly3(t, n)

woTD_poly4(t, n)

woTD_poly5(t, n)

woTD_poly6(t, n)

woTD_gam(t, n)

wD_poly3(t, n)

wD_poly4(t, n)

wD_poly5(t, n)

wD_poly6(t, n)

wD_gam(t, n)

wTD_poly3(t, n)

wTD_poly4(t, n)

wTD_poly5(t, n)

wTD_poly6(t, n)

wTD_gam(t, n)
}
\arguments{
\item{t}{Test statistics}

\item{n}{Sample size}
}
\value{
p-value from the ADF test statistic distributions.
}
\description{
Models for approximating p-value.
}
\details{
These are the functions based on polynomial regression and GAM to approximate
the p-value for ADF unit root tests of three types: without drift and trend
(woTD), with drift (wD) and with both drift and trend (wTD).
}
